+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- General
+  - Add rewrite-clj node to cursor-info.
+
 ## 2023.07.01-22.35.41
 
 - General